+++ R/civis_ml_utils.R
@@ -20,9 +20,9 @@ CIVIS_ML_CLASSIFIERS <- c("sparse_logistic",
                           "stacking_classifier")
 
 CIVIS_ML_TEMPLATE_IDS <- data.frame(
-  id = c(9112, 9113, 9968, 9969, 10582, 10583),
-  version = c(1.1, 1.1, 2.0, 2.0, 2.1, 2.1),
-  name = c("train", "predict", "train", "predict", "train", "predict"),
+  id = c(9112, 9113, 9968, 9969, 10582, 10583, 11219, 11220, 11221),
+  version = c(1.1, 1.1, 2.0, 2.0, 2.1, 2.1, 2.2, 2.2, 2.2),
+  name = c("train", "predict", "train", "predict", "train", "predict", "train", "predict", "register"),
   stringsAsFactors = FALSE
 )