+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - `write_civis_file` now has a defaults to the file path for the `name` argument instead of requiring it.
 
+### Fixed
+
+- `read_civis.numeric`, `write_civis.numeric`, `download_civis.numeric`, and `query_civis.numeric` now fail with a better error message for `NA` of type numeric.
+
 ## [1.2.0] - 2018-01-23
 
 ### Fixed