+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ### Changed
 
+- Update CivisML stacking documentation.
 - `write_civis_file` now defaults to the file path for the `name` argument instead of requiring it.
 
 ### Fixed