+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - `read_civis.numeric`, `write_civis.numeric`, `download_civis.numeric`, and `query_civis.numeric` now fail with a better error message for `NA` of type numeric.
 - The `hidden` argument of `write_civis` now works.
+- Remove new lines from parameter descriptions in generated docs.
 
 ## [1.2.0] - 2018-01-23