+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-## Unreleased
+## [1.4.0] - 2018-05-04
 
 ### Changed
 - Add v2.2 CivisML templates.
+- The DBI connectors are provided on an as-is basis and are no longer maintained.
 
+### Fixed
+- Fix CRAN check failures for dev DBI-test.
 
 ## [1.3.0] - 2018-03-30